% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment_OR.R
\name{All_motif_inst_OR}
\alias{All_motif_inst_OR}
\title{Motif instances enrichment all filters}
\usage{
All_motif_inst_OR(
  EM_list,
  Clinvar_path,
  iELM_insts,
  new_hits_ints,
  pepsite_res,
  prot_seqs,
  bench_data = NULL,
  min_dom = 5,
  adj_pval = 0.05,
  datasets_input = NULL,
  prot_only = F
)
}
\value{
A dataframe with odds ratio of each filter.
}
\description{
Motif instances enrichment all filters
}
\author{
Bishoy Wadie, Evangelia Petsalaki
}
