% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment_OR.R
\name{All_prot_OR}
\alias{All_prot_OR}
\title{Motif carrying protein enrichment all filters}
\usage{
All_prot_OR(
  EM_list,
  Clinvar_path,
  iELM_insts,
  new_hits_ints,
  pepsite_res,
  bench_data = NULL,
  min_dom = 5,
  adj_pval = 0.05,
  qslim_input = NULL,
  comparison = T,
  bg_possible = T,
  prot_only = T
)
}
\value{
A dataframe with odds ratio of each filter.
}
\description{
Motif carrying protein enrichment all filters
}
\author{
Bishoy Wadie, Evangelia Petsalaki
}
