% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment_OR.R
\name{Domain_Enrichment}
\alias{Domain_Enrichment}
\title{Domain enrichment of motif-carrying proteins}
\usage{
Domain_Enrichment(
  human_fasta,
  human_human_int,
  mot_carprots,
  cdhit_output = NULL,
  cdhit_outdir = NULL,
  pfam_doms_list = NULL,
  cdhit_path = NULL,
  cdhit_ident_thresh = 0.7,
  cdhit_no_cores = 1,
  cdhit_wsl = T,
  keep_cdhit_out = T
)
}
\value{
A list of non_redundant domains and their enrichment scores for each motif_carrying protein.
}
\description{
Domain enrichment of motif-carrying proteins
}
\author{
Bishoy Wadie, Evangelia Petsalaki
}
