% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chembl_Drug_Mapping.R
\name{H1_to_ChEMBL}
\alias{H1_to_ChEMBL}
\title{Maps Chembl Drugs to H1 proteins}
\usage{
H1_to_ChEMBL(
  slim_hits,
  clinvar_df,
  zooma_output = NULL,
  use_zooma = F,
  host_viral = T,
  chemble_source = NULL,
  Drug_inds = NULL,
  indirect_mapping_dist = 2
)
}
\value{
A list containing all Drugs targeting H1 proteins from ChEMBL database and the ones that share indication with ClinVar disease.
}
\description{
Maps ChEMBL small molecule compounds to H1 proteins and connects the drug indication with the ClinVar disease associated with the predicted motifs in H2 proteins.
}
\examples{
new_hits = parse_qslim()
Clinvar_mapped = Map_ClinVar(new_hits_int = new_hits)
H1_chembl = H1_to_ChEMBL(new_hits_int = new_hits, clinvar_df = Clinvar_mapped)
}
\author{
Bishoy Wadie, Evangelia Petsalaki
}
