% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{HV_motif_level_eval}
\alias{HV_motif_level_eval}
\title{Motif-level evaluation of predicted hits}
\usage{
HV_motif_level_eval(
  slim_hits,
  bench_data,
  bench_data_type = c("ELM", "PRMdb"),
  datasets_input = NULL,
  all_pfam_dom = NULL,
  prot_seqs = NULL
)
}
\value{
A data frame containing all the relevant motif-level performance metrics for each domain enrichment filter in addition to the non-filtered qslim output.
}
\description{
Calculates performance metrics for evaluating predicted hits against ELM true positive instances on the motif-level.
}
\details{
For motif-level enrichment, we simply cannot use a binary classification as we did for protein-level evaluation because in reality, predicted motifs are partially correct to some extent as they might contain true-positive residues in a given sequence stretch, and therefore we used a re-implemented version of the evaluation protocol proposed in \href{https://academic.oup.com/nar/article/45/W1/W470/3782606}{Prytuliak et al. 2017} instead of binary classification, where we computed the common performance metrics (Recall, precision F1, etc .. ) both residue-wise and site-wise given that the motif-carrying proteins are also found in the ELM benchmarking dataset. So this analysis was not performed on proteins not reported in the ELM dataset.
}
\examples{
motif_eval_metrics = HV_motif_level_eval()
}
\references{
Prytuliak, Roman, et al. "HH-MOTiF: de novo detection of short linear motifs in proteins by Hidden Markov Model comparisons." \emph{Nucleic acids research} 45.W1 (2017): W470-W477.
}
\author{
Bishoy Wadie, Evangelia Petsalaki
}
