% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{HV_prot_dom_int_eval}
\alias{HV_prot_dom_int_eval}
\title{Evaluation of protein-domain interactions}
\usage{
HV_prot_dom_int_eval()
}
\value{
A data frame containing all the relevant protein-domain interactions' performance metrics for each domain enrichment filter.
}
\description{
Calculates performance metrics for evaluating protein-domain interactions between the motif-carrying protein and the enriched domain against ELM interactions.
}
\details{
For evaluating protein-domain interactions we measured the enrichment of true-positive interactions between a given motif-carrying protein and its associated domains as reported in the ELM interaction dataset. As in the motif-level evaluation, this analysis was performed only on the motif-carrying proteins reported in the ELM interactions dataset, where true-positives represents the number of correctly associated domains for a given motif-carrying protein and then summed over all motif-carrying proteins in the predicted dataset.
}
\examples{
ProtDom_int_eval_metrics = HV_prot_dom_int_eval()
}
\author{
Bishoy Wadie, Evangelia Petsalaki
}
