% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{HV_prot_level_eval}
\alias{HV_prot_level_eval}
\title{Protein-level evaluation of predicted hits}
\usage{
HV_prot_level_eval()
}
\value{
A data frame containing all the relevant protein-level performance metrics for each domain enrichment filter in addition to the non-filtered qslim output.
}
\description{
Calculates performance metrics for evaluating predicted hits against ELM true positive instances on the protein-level.
}
\details{
For protein-level enrichment, we measured the enrichment of true-positives in our predicted dataset using a one-tailed fisher-exact test, where the odds ratio represents the magnitude of the enrichment. True-positives are the number of motif-carrying proteins present in both the predicted dataset and the ELM dataset regardless of whether the predicted protein has the right motif or found in the right location
}
\examples{
prot_eval_metrics = HV_prot_level_eval()
}
\author{
Bishoy Wadie, Evangelia Petsalaki
}
