% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVSlimPred-package.R
\docType{package}
\name{HVSlimPred-package}
\alias{HVSlimPred}
\alias{HVSlimPred-package}
\title{HVSlimPred: Identification of novel functional linear motifs using the host-viral protein interaction network and the principle of convergent evolution}
\description{
This package complements the performance evaluation analysis for the manuscript entitled "Identifying novel functional linear motifs using the host-viral protein interaction network and the principle of convergent evolution".
}
\keyword{internal}
