% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iELm_HMM.R
\name{iELM_HMM}
\alias{iELM_HMM}
\title{iELM HMM analysis on H1 proteins}
\usage{
iELM_HMM(
  path_HMM_pfam_domtblouts = NULL,
  path_HMM_iELM_domtblouts = NULL,
  pct_coverage = 0.8,
  ELM_renamed_classes = NULL
)
}
\value{
Combined output from HMMER against both iELM and pfam HMMs for all H1 proteins.
}
\description{
Preprocess and evaluate identified motif-binding domains in H1 proteins using both iELM and PFAM HMMs.
}
\examples{
HMM_results = iELM_HMM()
}
\references{
Weatheritt, Robert J., et al. "iELM—a web server to explore short linear motif-mediated interactions." \emph{Nucleic acids research} 40.W1 (2012): W364-W369.
}
\author{
Bishoy Wadie, Evangelia Petsalaki
}
