% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clinvar_and_PTMs.R
\name{Map_ClinVar}
\alias{Map_ClinVar}
\title{Map predicted motifs to ClinVar variants}
\usage{
Map_ClinVar(
  slim_hits,
  Clinvar_variant_summary = NULL,
  pathogenic_only = T,
  qslim_noAA_gncoords = NULL
)
}
\value{
A data frame containing ClinVar variants mapped to all predicted motifs.
}
\description{
Maps the predicted motifs to disease variants retrieved from ClinVar.
}
\examples{
PTM_mapping = Map_PTM()
}
\author{
Bishoy Wadie, Evangelia Petsalaki
}
