% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clinvar_and_PTMs.R
\name{Map_PTM}
\alias{Map_PTM}
\title{Map predicted motifs to PTMs}
\usage{
Map_PTM(slim_hits, all_pfam_dom = NULL, uniprot_mod_res = NULL)
}
\value{
A data frame containing PTMs mapped to all predicted motifs.
}
\description{
Maps the predicted motifs to PTMs retrieved from uniprot.
}
\examples{
PTM_mapping = Map_PTM()
}
\author{
Bishoy Wadie, Evangelia Petsalaki
}
