% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qslim.R
\name{parse_qslim}
\alias{parse_qslim}
\title{Parse qslim output per host-viral interaction data}
\format{
\itemize{
\item \emph{Dataset}: Uniprot IDs of Host-viral interaction denoted as Viral-prot_Human-prot
\item \emph{Pattern}: Motif's regular expression pattern
\item \emph{uniprot}: Uniprot ID of motif-carrying protein
\item \emph{Org}: Organism info about the motif-carrying protein
\item \emph{Start_Pos}: First position of the predicted motif
\item \emph{End_Pos}: Last position of the predicted motif
\item \emph{Desc}: Uniprot description of the motif-carrying protein
\item \emph{Id}: ELM Identifier
\item \emph{Sim_rank}: Ranking of CompariMotif relationships. A value of 9 corresponds to Exact match, while a value of 1 corresponds to Complex Overlap (Check CompariMotif \href{http://bioware.ucd.ie/~compass/biowareweb/Server_pages/help/slimfinder/comparimotif_help.html#relationships}{Website} for more details)
\item \emph{Score}: CompariMotif heuristic similarity score, defined as the product of Matched positions and Normalized Information content (Check CompariMotif \href{https://academic.oup.com/bioinformatics/article/24/10/1307/177233}{paper} for more details)
\item \emph{Accession}: ELM Instance Accession
\item \emph{start}: First position of the ELM motif
\item \emph{end}: Last position of the ELM motif
\item \emph{Logic}: ELM instance logic as reported by ELM database.
\item \emph{motif_distance}: Distance between predicted and ELM motif on the same protein
\item \emph{motif_overlap}: Overlap proportion between predicted and ELM motif on the same protein
}
}
\usage{
parse_qslim(
  main_output = NULL,
  occ_out = NULL,
  compari_output = NULL,
  ELM_input = NULL,
  preprocess_compari = T,
  select_top_rank = F,
  IC_cutoff = 3,
  remove_motif_variants = T
)
}
\value{
A data frame containing all the predicted hits annotated with corresponding ELM instances and CompariMotif similarity scores.
}
\description{
Parses the raw combined qslim output per host-viral interaction dataset. Then it adds the corresponding ELM instances along with the CompariMotif similarity between the predicted regular expression motif patterns and that of ELM classes. It also includes the distance between the predicted motif and that of an ELM instance, only if the motif-carrying protein and ELM class match.
}
\examples{
pred_hits = parse_qslim()
}
\author{
Bishoy Wadie, Evangelia Petsalaki
}
