% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chembl_Drug_Mapping.R
\name{send_to_zooma}
\alias{send_to_zooma}
\title{Prepare Input to Zooma}
\usage{
send_to_zooma(clinvar_mapped, output_all_disease_info = F)
}
\value{
A character vector of Disease names
}
\description{
Send the output to Zooma as query
}
\author{
Bishoy Wadie, Evangelia Petsalaki
}
